Refactor Remove InteractorResponse from notes
