+++ backend/src/notes/interactors/index.ts
@@ -1,2 +1,1 @@
-export * from './InteractorResponse';
 export * from './NoteInteractor';
